Move CubeSide to separate component
